refactor: changes requested for the pr
