+++ packages/ssi-react/src/styles/components/components/SSIText/index.ts
